feat(LottoChecker): UI 및 컨트롤러 추가
- InputView 추가
- OutputView 추가
- Lotto 추가
- LottoBuyer, LottoResult에 spentMoney long형 변경 및 멤버변수명 수정
- LottoRank print용 getter 추가
- LottoNumber print용 toString 추가 및 수정

+++ src/main/java/lotto/domain/LottoBuyer.java
@@ -13,9 +13,9 @@ public class LottoBuyer {
 			.boxed()
 			.collect(Collectors.toList());
 
-	public List<LottoNumber> buyLottoNumbers(int money) {
-		int count = money / PRIZE;
-		return getLottoNumbers(count);
+	public List<LottoNumber> buyLottoNumbers(long money) {
+		long count = money / PRIZE;
+		return getLottoNumbers((int) count);
 	}
 
 	private List<LottoNumber> getLottoNumbers(int count) {
